requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pydantic[email]==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
supabase==1.0.3
crawl4ai==0.2.76
beautifulsoup4==4.12.2
slowapi==0.1.8
httpx==0.25.0
pytest==8.3.3
pytest-asyncio==0.23.5
pytest-mock==3.12.0